Fix test that robo-tools decided we shouldn't use assertEquals on.

+++ tests/Controllers/Admin/UserSettingsTest.php
@@ -90,6 +90,6 @@ final class UserSettingsTest extends TestCase
         $this->assertTrue(setting('Auth.sessionConfig')['allowRemembering']);
         $this->assertSame(TwoFactorAuthEmail::class, setting('Auth.actions')['login']);
         $this->assertSame(EmailActivator::class, setting('Auth.actions')['register']);
-        $this->assertSame(2 * WEEK, setting('Auth.sessionConfig')['rememberLength']);
+        $this->assertSame(2 * WEEK, (int)setting('Auth.sessionConfig')['rememberLength']);
     }
 }
